feat: prevent from deleting own sessions during managing sessions

+++ src/api/admin_session_controller.rs
@@ -2,8 +2,9 @@ use crate::api::dto::Pagination;
 use crate::api::dto::{MessageResponse, SessionListResponse};
 use crate::api::server_state::ServerState;
 use crate::application::auth_service::AuthStrategy;
+use crate::domain::jwt::UserDTO;
 use crate::domain::session::Session;
-use axum::extract::{Path, Query, State};
+use axum::extract::{Extension, Path, Query, State};
 use axum::http::StatusCode;
 use axum::response::IntoResponse;
 use axum::Json;
@@ -105,6 +106,7 @@ pub async fn get_session(
 )]
 pub async fn delete_session(
     State(state): State<ServerState>,
+    Extension(current_user): Extension<UserDTO>,
     Path(id): Path<Uuid>,
 ) -> impl IntoResponse {
     if state.config.auth_strategy() == AuthStrategy::Stateless {
@@ -117,14 +119,29 @@ pub async fn delete_session(
             .into_response();
     }
 
-    match state.session_repository.delete(&id).await {
-        Ok(_) => (
-            StatusCode::OK,
-            Json(MessageResponse {
-                message: "Session deleted successfully".to_string(),
-            }),
-        )
-            .into_response(),
+    match state.session_repository.get_by_id(&id).await {
+        Ok(session) => {
+            if session.user_id == current_user.id {
+                return (
+                    StatusCode::BAD_REQUEST,
+                    Json(MessageResponse {
+                        message: "Cannot delete your own session".to_string(),
+                    }),
+                )
+                    .into_response();
+            }
+
+            match state.session_repository.delete(&id).await {
+                Ok(_) => (
+                    StatusCode::OK,
+                    Json(MessageResponse {
+                        message: "Session deleted successfully".to_string(),
+                    }),
+                )
+                    .into_response(),
+                Err(e) => e.into_response(),
+            }
+        }
         Err(e) => e.into_response(),
     }
 }
@@ -143,6 +160,7 @@ pub async fn delete_session(
 )]
 pub async fn delete_all_user_sessions(
     State(state): State<ServerState>,
+    Extension(current_user): Extension<UserDTO>,
     Path(user_id): Path<Uuid>,
 ) -> impl IntoResponse {
     if state.config.auth_strategy() == AuthStrategy::Stateless {
@@ -150,6 +168,16 @@ pub async fn delete_all_user_sessions(
             StatusCode::BAD_REQUEST,
             Json(MessageResponse {
                 message: "Action not supported in stateless this strategy".to_string(),
+            }),
+        )
+            .into_response();
+    }
+
+    if user_id == current_user.id {
+        return (
+            StatusCode::BAD_REQUEST,
+            Json(MessageResponse {
+                message: "Cannot delete your own sessions".to_string(),
             }),
         )
             .into_response();
